Refactor `objects_to_docs` and `rerank_docs` methods Refactor vector database clearing after superceding web-retrieval

+++ python-ai-server/RAG/web_weaviate.py
@@ -57,10 +57,10 @@ class WebWeaviate(VectorStore):
     def delete(self):
         response = self.collection.query.fetch_objects(limit=FETCHING_LIMIT, return_properties=[])
         ids = [o.uuid for o in response.objects]
-
-        self.collection.data.delete_many(
-            where=wvc.query.Filter.by_id().contains_any(ids)
-        )
+        if(ids):
+            self.collection.data.delete_many(
+                where=wvc.query.Filter.by_id().contains_any(ids)
+            )
     # -------------------------------------------------- #
         
     # Response to documents
